API: Move api/config/fonts.py to api/fonts.py
That's because it'll be easier for third-party developers to find that
module this way.
