+++ spyder/api/fonts.py
