+++ spyder/plugins/help/plugin.py
@@ -15,8 +15,7 @@ import os
 from qtpy.QtCore import Signal
 
 # Local imports
-from spyder.api.config.fonts import SpyderFontType
-from spyder.api.exceptions import SpyderAPIError
+from spyder.api.fonts import SpyderFontType
 from spyder.api.plugins import Plugins, SpyderDockablePlugin
 from spyder.api.plugin_registration.decorators import (
     on_plugin_available, on_plugin_teardown)
