+++ spyder/plugins/history/plugin.py
@@ -12,7 +12,7 @@ Console History Plugin.
 from qtpy.QtCore import Signal
 
 # Local imports
-from spyder.api.config.fonts import SpyderFontType
+from spyder.api.fonts import SpyderFontType
 from spyder.api.plugins import Plugins, SpyderDockablePlugin
 from spyder.api.plugin_registration.decorators import (
     on_plugin_available, on_plugin_teardown)
