+++ spyder/plugins/ipythonconsole/plugin.py
@@ -16,7 +16,7 @@ from typing import List
 from qtpy.QtCore import Signal, Slot
 
 # Local imports
-from spyder.api.config.fonts import SpyderFontType
+from spyder.api.fonts import SpyderFontType
 from spyder.api.plugins import Plugins, SpyderDockablePlugin
 from spyder.api.plugin_registration.decorators import (
     on_plugin_available, on_plugin_teardown)
