+++ spyder/plugins/remoteclient/widgets/connectiondialog.py
@@ -29,8 +29,8 @@ from qtpy.QtWidgets import (
 )
 
 # Local imports
+from spyder.api.fonts import SpyderFontType, SpyderFontsMixin
 from spyder.api.translations import _
-from spyder.api.config.fonts import SpyderFontType, SpyderFontsMixin
 from spyder.api.utils import get_class_values
 from spyder.api.widgets.dialogs import SpyderDialogButtonBox
 from spyder.plugins.remoteclient.api.protocol import (
